// backend/constantes/enum.ts
export enum NodeType {
    SERVER = 'server',
    NODE = 'node',
    CLIENT = 'client'
}

export enum LogLevelColors {
    DEBUG = '\x1b[32m',
    INFO = '\x1b[36m',
    WARN = '\x1b[33m',
    ERROR = '\x1b[31m',
    END = '\x1b[0m'
}

export enum LogLevel {
    DEBUG = 'DEBUG',
    INFO = 'INFO',
    WARN = 'WARN',
    ERROR = 'ERROR',
}

export enum LogLevelValues {
    DEBUG = 20,
    INFO = 30,
    WARN = 40,
    ERROR = 50
}